refactor: move FLN_LEVELS_LIST and FLNLevelReferenceModal out of RoleDashboards.tsx
Closes #138. Extracted into frontend/src/components/dashboards/FLNLevelReference.tsx,
re-exported from RoleDashboards.tsx for existing importers (TeacherDashboard,
VolunteerDashboard, ContentPanel). No logic changes — copied as-is.

Verified: tsc --noEmit and vite build both clean; live click-through confirms
the modal opens, search/class-filter work, and closes cleanly with no console
errors.
